make isPrintable available to users
