+++ apf/apf.cc
@@ -17,6 +17,8 @@
 #include "apfArrayData.h"
 #include "apfTagData.h"
 #include "apfUserData.h"
+#include "apfVtk.h"
+#include "apfNumberingClass.h"
 #include <cstdio>
 #include <cstdlib>
 #include <pcu_util.h>
@@ -436,7 +438,12 @@ void sharedReduction(Field* f, Sharing* shr, bool delete_shr,
   reduceFieldData(f->getData(), shr, delete_shr, sum);
 }
 
-
+bool isPrintable(Field* f)
+{
+  // cast to FieldBase and call the other method
+  FieldBase* f2 = f;
+  return isPrintable(f2);
+}
 
 void fail(const char* why)
 {
